Separate index implementation from IndexedBucket
