+++ packages/storage/src/lib.rs
@@ -1,5 +1,6 @@
 mod bucket;
 mod indexed_bucket;
+mod indexes;
 mod length_prefixed;
 mod namespace_helpers;
 mod prefixed_storage;
@@ -11,7 +12,7 @@ mod typed;
 
 pub use bucket::{bucket, bucket_read, Bucket, ReadonlyBucket};
 #[cfg(feature = "iterator")]
-pub use indexed_bucket::{Core, IndexedBucket, MultiIndex, UniqueIndex};
+pub use indexed_bucket::IndexedBucket;
 pub use length_prefixed::{to_length_prefixed, to_length_prefixed_nested};
 pub use prefixed_storage::{prefixed, prefixed_read, PrefixedStorage, ReadonlyPrefixedStorage};
 pub use sequence::{currval, nextval, sequence};
